+++ src/openpecha/alignment/parsers/plaintext.py
@@ -9,6 +9,7 @@ from openpecha.alignment import Alignment, AlignmentMetaData
 from openpecha.alignment.metadata import AlignmentRelationEnum
 from openpecha.config import _mkdir
 from openpecha.ids import get_initial_pecha_id, get_uuid
+from openpecha.pecha import save_stam
 from openpecha.pecha.layer import (
     LayerCollectionEnum,
     LayerEnum,
@@ -270,31 +271,3 @@ def split_text_into_lines(text: str) -> List[str]:
     """
     lines = text.split("\n")
     return [line + "\n" if i < len(lines) - 1 else line for i, line in enumerate(lines)]
-
-
-def save_stam(ann_store: AnnotationStore, ann_store_path: Path) -> Path:
-    """
-    Save the annotation store to a file.
-    """
-    ann_store_path.parent.mkdir(parents=True, exist_ok=True)
-
-    ann_json_str = ann_store.to_json_string()
-    ann_json_dict = convert_absolute_to_relative_path(ann_json_str, ann_store_path)
-    with open(ann_store_path, "w", encoding="utf-8") as f:
-        f.write(json.dumps(ann_json_dict, indent=2, ensure_ascii=False))
-
-    return ann_store_path
-
-
-def convert_absolute_to_relative_path(json_string: str, ann_store_path: Path):
-    """
-    convert the absolute to relative path for base file in json string of annotation store
-    """
-    json_object = json.loads(json_string)
-    for resource in json_object["resources"]:
-        original_path = Path(resource["@include"])
-        if ann_store_path.name == "metadata.json":
-            resource["@include"] = f"base/{original_path.name}"
-        else:
-            resource["@include"] = f"../base/{original_path.name}"
-    return json_object
